+++ client/src/App.tsx
@@ -4,8 +4,8 @@ import './assets/stylesheets/colors.css';
 import './assets/stylesheets/buttons.css';
 import './assets/stylesheets/borders.css';
 import {BrowserRouter as Router, Route, Routes} from 'react-router-dom';
-import ADDRESSES from './constants/addresses/addresses';
-import { PrimaryLayout } from './components/layouts/MainLayout';
+import { ADDRESSES } from '@/constants';
+import { PrimaryLayout } from '@/components/layouts/MainLayout';
 import LandingPage from './features/landing/LandingPage';
 import Home from "./features/home/Home";
 import StudentProjects from './features/project-list/StudentProjects';
