+++ client/src/features/project-details/components/ProjectInformation.tsx
@@ -1,10 +1,11 @@
 import { createContext, useContext, useEffect, useMemo, useState } from "react";
 import { Color, OnHover, Opacity } from "@/types";
 import TagRow from "@components/shared/TagRow";
-import { ICONS } from "@constants/project_data/tags";
-import { PROJECT_DESCRIPTION } from "@constants/defaults";
+import { ICONS, DEFAULTS } from "@/constants";
 import { IProjectTags, ProjectDataService } from "@utils/data/ProjectDataService";
 import { AllTagsContext, ProjectContext, UserContext } from "./ProjectDetailsProvider";
+
+const PROJECT_DESCRIPTION = DEFAULTS.PROJECT_DESCRIPTION;
 
 const DevIconContext = createContext<string>('');
 const ProjectTagsContext = createContext<IProjectTags | undefined>(undefined);
